move manageservice tests to shared folder, extract shared mocks
