+++ Composer/packages/client/__mocks__/@azure/arm-resources.tsx
@@ -1,0 +1,19 @@
+let mock = jest.createMockFromModule('@azure/arm-resources');
+
+mock.ResourceManagementClient = () => {
+  return {
+    resourceGroups: {
+      list: async () => {
+        return [
+          {
+            id: 'mockedGroup',
+            name: 'mockedGroup',
+            region: 'westus',
+          },
+        ];
+      },
+    },
+  };
+};
+
+module.exports = mock;
