+++ Composer/packages/client/__mocks__/@azure/ms-rest-js.tsx
@@ -1,0 +1,4 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+module.exports = jest.createMockFromModule('@azure/ms-rest-js');
